Refactors EndpointToSttpClientExtensions * The only thing that can be found there is related to WebSockets, hence the renaming and extending the trait by WebSocketEndpointToSttpClient, not the base one
