+++ client/sttp-client4/src/main/scalajvm/sttp/tapir/client/sttp4/WebSocketEndpointToSttpClientExtensions.scala
@@ -1,8 +1,8 @@
-package sttp.tapir.client.sttp
+package sttp.tapir.client.sttp4
 
 import sttp.model.StatusCode
 
-private[sttp] trait EndpointToSttpClientExtensions { this: EndpointToSttpClient[_] =>
+private[sttp4] trait WebSocketEndpointToSttpClientExtensions {
 
   /** This needs to be platform-specific due to #2663, as on JS we don't get access to the 101 status code. */
   val webSocketSuccessStatusCode: StatusCode = StatusCode.SwitchingProtocols
